debugging update spot

+++ backend/db/migrations/20241105184529-create-spot.js
@@ -58,7 +58,7 @@ module.exports = {
           allowNull: false,
         },
         price: {
-          type: Sequelize.DECIMAL(10, 2),
+          type: Sequelize.FLOAT(10, 2),
           allowNull: false,
         },
         createdAt: {
